AI agents/camelAI/requirements.txt
flask
requests
camel-ai
python-dotenv
beautifulsoup4
